requirements.txt
django
djangorestframework
hdwallet
requests
